backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    ./scripts/generate_requirements.sh upgrade
agno==1.4.6
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl-cffi==0.12.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
duckduckgo-search==8.1.1
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
frozendict==2.4.6
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
proto-plus==1.26.1
pyparsing==3.2.3
uritemplate==4.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multitasking==0.0.12
numpy==2.3.2
openai==1.97.1
pandas==2.3.1
peewee==3.18.2
pgvector==0.4.1
platformdirs==4.3.8
primp==0.15.0
protobuf==6.31.1
psycopg==3.2.9
psycopg-binary==3.2.9
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
rsa==4.9
cachetools==5.5.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
starlette==0.47.2
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yfinance==0.2.65
youtube-transcript-api==1.2.1
yt-dlp==2024.11.18
ffmpeg-python==0.2.0
pillow==10.4.0
opencv-python==4.10.0.84
moviepy==1.0.3
openai-whisper==20231117
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.2.1
chromadb==0.5.20
azure-storage-blob==12.24.0
playwright==1.48.0
